requirements.txt
absl-py==0.11.0
arpa==0.1.0b4
astor==0.8.1
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
deepdiff==5.0.2
farmi==3.1.0
gast==0.4.0
gensim==3.8.3
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.9.0
idna==2.10
importlib-metadata==3.1.1
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
msgpack==1.0.0
msgpack-numpy==0.4.7.1
numpy==1.19.4
ordered-set==4.0.2
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
progress==1.5
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
smart-open==4.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.8.0
tensorflow-text==0.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
